Moved functions get_all_derived_variables to _derive module
