+++ esmvaltool/preprocessor/_derive/__init__.py
@@ -1,7 +1,9 @@
 """Automatically derive variables."""
 
 
+import importlib
 import logging
+import os
 
 import iris
 import yaml
@@ -9,6 +11,8 @@ import yaml
 from ._derived_variable_base import DerivedVariableBase
 
 logger = logging.getLogger(__name__)
+
+ALL_DERIVED_VARIABLES = {}
 
 
 def get_required(short_name, field=None):
@@ -79,3 +83,32 @@ def derive(cubes, variable):
     cube.attributes['metadata'] = yaml.safe_dump(variable)
 
     return cube
+
+def get_all_derived_variables():
+    """Get all possible derived variables.
+
+    Returns
+    -------
+    dict
+        All derived variables with `short_name` (keys) and the associated
+        python classes (values).
+
+    """
+    if ALL_DERIVED_VARIABLES:
+        return ALL_DERIVED_VARIABLES
+    current_path = os.path.dirname(os.path.realpath(__file__))
+    for var_file in os.listdir(current_path):
+        var_name = os.path.splitext(var_file)[0]
+        try:
+            var_module = importlib.import_module(
+                'esmvaltool.preprocessor._derive.{}'.format(var_name))
+            try:
+                derived_var = getattr(var_module,
+                                      'DerivedVariable')(var_name)
+                ALL_DERIVED_VARIABLES[var_name] = derived_var
+            except AttributeError:
+                pass
+        except ImportError:
+            pass
+
+    return ALL_DERIVED_VARIABLES
