+++ esmvaltool/preprocessor/_derive/_derived_variable_base.py
@@ -3,12 +3,8 @@
 
 import importlib
 import logging
-import os
 
 logger = logging.getLogger(__name__)
-
-
-ALL_DERIVED_VARIABLES = {}
 
 
 class DerivedVariableBase(object):
@@ -110,33 +106,3 @@ class DerivedVariableBase(object):
                            "preprocessor/_derive/ for variable derivation",
                            short_name)
         return derived_var
-
-    @staticmethod
-    def get_all_derived_variables():
-        """Get all possible derived variables.
-
-        Returns
-        -------
-        dict
-            All derived variables with `short_name` (keys) and the associated
-            python classes (values).
-
-        """
-        if ALL_DERIVED_VARIABLES:
-            return ALL_DERIVED_VARIABLES
-        current_path = os.path.dirname(os.path.realpath(__file__))
-        for var_file in os.listdir(current_path):
-            var_name = os.path.splitext(var_file)[0]
-            try:
-                var_module = importlib.import_module(
-                    'esmvaltool.preprocessor._derive.{}'.format(var_name))
-                try:
-                    derived_var = getattr(var_module,
-                                          'DerivedVariable')(var_name)
-                    ALL_DERIVED_VARIABLES[var_name] = derived_var
-                except AttributeError:
-                    pass
-            except ImportError:
-                pass
-
-        return ALL_DERIVED_VARIABLES
